Split model/index.js into smaller files
